Make single command
