+++ .gitignore
@@ -1,4 +1,9 @@
+.vscode
 .idea
+.DS_Store
+
 target
 Cargo.lock
 perf.data*
+
+/bin
