+++ Makefile
@@ -6,13 +6,16 @@ ifeq ($(VERSION),)
 endif
 
 clean:
+	rm -rf $(BIN_DIR)
 	cargo clean
 
 format:
 	cargo fmt
 
 build:
+	mkdir -p $(BIN_DIR)
 	cargo build --release
+	cp -p ./target/release/lindera $(BIN_DIR)
 
 test:
 	cargo test
