+++ src/core/dict.rs
@@ -1,8 +1,8 @@
-use std::io::{self, Read, BufRead, BufReader};
-use std::fs::File;
-use std::u32;
-use std::str::FromStr;
 use fst;
+use std::fs::File;
+use std::io::{self, BufRead, BufReader, Read};
+use std::str::FromStr;
+use std::u32;
 
 const DICTIONARY_DATA: &'static [u8] = include_bytes!("../../dict/dict.fst");
 
@@ -11,11 +11,9 @@ pub struct Dict {
 }
 
 impl Dict {
-
     pub fn load_default() -> Dict {
         Dict {
             fst: fst::raw::from_static_slice(DICTIONARY_DATA),
         }
     }
 }
-
